+++ src/common/utils/filesUtils.ts
@@ -221,24 +221,37 @@ export class FilesExporter {
     this.totalSoqlRequests++;
     const contentVersions = await bulkQuery(contentVersionSoql, this.conn);
 
+    // ContentDocument object can be linked to multiple other objects even with same type (for example: same attachment can be linked to multiple EmailMessage objects).
+    // Because of this when we fetch ContentVersion for ContentDocument it can return less results than there is ContentDocumentLink objects to link.
+    // To fix this we create a list of ContentVersion and ContentDocumentLink pairs. 
+    // This way we have multiple pairs and we will download ContentVersion objects for each linked object.
+    const versionsAndLinks = []
+    contentVersions.records.forEach(contentVersion => {
+      contentDocumentLinks.records.forEach(contentDocumentLink => {
+        if (contentDocumentLink.ContentDocumentId === contentVersion.ContentDocumentId) {
+          versionsAndLinks.push({
+            contentVersion: contentVersion,
+            contentDocumentLink: contentDocumentLink,
+          });
+        }
+      });
+    });
+
     // Download files
     await PromisePool.withConcurrency(5)
-      .for(contentVersions.records)
-      .process(async (contentVersion: any) => {
+      .for(versionsAndLinks)
+      .process(async (versionAndLink: any) => {
         try {
-          await this.downloadContentVersionFile(contentVersion, records, contentDocumentLinks.records);
+          await this.downloadContentVersionFile(versionAndLink.contentVersion, records, versionAndLink.contentDocumentLink);
         } catch (e) {
           this.filesErrors++;
-          uxLog(this, c.red("Download file error: " + contentVersion.Title + "\n" + e));
+          uxLog(this, c.red("Download file error: " + versionAndLink.contentVersion.Title + "\n" + e));
         }
       });
   }
 
-  private async downloadContentVersionFile(contentVersion, records, contentDocumentLinks) {
+  private async downloadContentVersionFile(contentVersion, records, contentDocumentLink) {
     // Retrieve initial record to build output files folder name
-    const contentDocumentLink = contentDocumentLinks.filter(
-      (contentDocumentLink) => contentDocumentLink.ContentDocumentId === contentVersion.ContentDocumentId,
-    )[0];
     const parentRecord = records.filter((record) => record.Id === contentDocumentLink.LinkedEntityId)[0];
     // Build record output files folder (if folder name contains slashes or antislashes, replace them by spaces)
     const parentFolderName = (parentRecord[this.dtl.outputFolderNameField] || parentRecord.Id).replace(/[/\\?%*:|"<>]/g, "-");